Provide some first test cases
The idea is to have Java code generate these test cases, but that generation
is triggered manually (via “npm run gentest”) and the result is stored in
the repository.  That way running the tests is independent from having
access to a Java environment, and changes to the serialized format will
become more readily apparent.  Further tests expected to be added soon.